// src/app/enums/static.enums.ts
import { Champions } from './../interfaces/champion.interface';
export enum ePosition {
  TOP = 'TOP',
  MID = 'MID',
  JUNGLE = 'JUNGLE',
  BOT = 'BOT',
  SUPPORT = 'SUPPORT'
}

export enum eTeams {
  BLUE,
  RED
}

export enum eMapPosition {
  'TOP',
  'MID',
  'JUNGLE',
  'BOT',
  'BASE'
}

export enum eChampionNames {
  STRANGE = 'Dr. Strange',
  SPIDER = 'Spiderman',
  IRON = 'Iron Man',
  AMERICA = 'Captain America',
  HULK = 'Hulk',
  THOR = 'Thor',
  HAWKEYE = 'Hawkeye',
  WIDOW = 'Black Widow',
  MARVEL = 'Captain Marvel',
  ANTMAN = 'Antman',
}